Move back testutils to internal package
