+++ engines/router/missionctl/fiberapi/fan_in_test.go
@@ -15,7 +15,7 @@ import (
 	runnerV1 "github.com/caraml-dev/turing/engines/experiment/plugin/inproc/runner"
 	"github.com/caraml-dev/turing/engines/experiment/runner"
 	"github.com/caraml-dev/turing/engines/router/missionctl/experiment"
-	tfu "github.com/caraml-dev/turing/engines/router/missionctl/fiberapi/testutils"
+	tfu "github.com/caraml-dev/turing/engines/router/missionctl/fiberapi/internal/testutils"
 	tu "github.com/caraml-dev/turing/engines/router/missionctl/internal/testutils"
 	"github.com/gojek/fiber"
 	fiberHttp "github.com/gojek/fiber/http"
